+++ plugins/de.tudresden.slr.questionnaire/src/de/tudresden/slr/questionnaire/wizard/BadUserInputException.java
@@ -2,7 +2,7 @@ package de.tudresden.slr.questionnaire.wizard;
 
 @SuppressWarnings("serial")
 public class BadUserInputException extends Exception {
-	public BadUserInputException(String message) {
-		super(message);
-	}
+    public BadUserInputException(String message) {
+        super(message);
+    }
 }
